server/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Werkzeug==3.1.3
zipp==3.21.0
tensorflow==2.13.0
tensorflow-hub==0.13.0
Pillow==9.5.0
